mcp-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
chromadb==0.5.23
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0
mcp[cli]==1.3.0
PyYAML==6.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pypdf==5.1.0
python-docx==1.1.2
